--- rust/src/write/handlers/mod.rs
@@ -1,3 +1,0 @@
-//! handlers for processing messages in different formats in DeltaWriter
-pub mod json;
-pub mod record_batch;
