+++ rust/src/write/stats.rs
@@ -347,3 +347,321 @@ fn arrow_array_from_bytes(
 
     make_array(data)
 }
+
+#[cfg(test)]
+mod tests {
+    use super::json::record_batch_from_message;
+    use super::*;
+    use crate::{
+        action::{ColumnCountStat, ColumnValueStat},
+        DeltaTable, DeltaTableError,
+    };
+    use lazy_static::lazy_static;
+    use serde_json::{json, Value};
+    use std::collections::HashMap;
+    use std::path::Path;
+
+    #[tokio::test]
+    async fn delta_stats_test() {
+        let temp_dir = tempfile::tempdir().unwrap();
+        let table_path = temp_dir.path();
+        create_temp_table(table_path);
+
+        let table = load_table(table_path.to_str().unwrap(), HashMap::new())
+            .await
+            .unwrap();
+
+        let mut writer = DataWriter::for_table(&table, HashMap::new()).unwrap();
+
+        let arrow_schema = writer.arrow_schema();
+        let batch = record_batch_from_message(arrow_schema, JSON_ROWS.clone().as_ref()).unwrap();
+
+        writer.write(vec![batch]).await.unwrap();
+        let add = writer.write_parquet_files(&table.table_uri).await.unwrap();
+        assert_eq!(add.len(), 1);
+        let stats = add[0].get_stats().unwrap().unwrap();
+
+        let min_max_keys = vec!["meta", "some_int", "some_string", "some_bool"];
+        let mut null_count_keys = vec!["some_list", "some_nested_list"];
+        null_count_keys.extend_from_slice(min_max_keys.as_slice());
+
+        assert_eq!(min_max_keys.len(), stats.min_values.len());
+        assert_eq!(min_max_keys.len(), stats.max_values.len());
+        assert_eq!(null_count_keys.len(), stats.null_count.len());
+
+        // assert on min values
+        for (k, v) in stats.min_values.iter() {
+            match (k.as_str(), v) {
+                ("meta", ColumnValueStat::Column(map)) => {
+                    assert_eq!(2, map.len());
+
+                    let kafka = map.get("kafka").unwrap().as_column().unwrap();
+                    assert_eq!(3, kafka.len());
+                    let partition = kafka.get("partition").unwrap().as_value().unwrap();
+                    assert_eq!(0, partition.as_i64().unwrap());
+
+                    let producer = map.get("producer").unwrap().as_column().unwrap();
+                    assert_eq!(1, producer.len());
+                    let timestamp = producer.get("timestamp").unwrap().as_value().unwrap();
+                    assert_eq!("2021-06-22", timestamp.as_str().unwrap());
+                }
+                ("some_int", ColumnValueStat::Value(v)) => assert_eq!(302, v.as_i64().unwrap()),
+                ("some_bool", ColumnValueStat::Value(v)) => assert_eq!(false, v.as_bool().unwrap()),
+                ("some_string", ColumnValueStat::Value(v)) => {
+                    assert_eq!("GET", v.as_str().unwrap())
+                }
+                ("date", ColumnValueStat::Value(v)) => {
+                    assert_eq!("2021-06-22", v.as_str().unwrap())
+                }
+                _ => assert!(false, "Key should not be present"),
+            }
+        }
+
+        // assert on max values
+        for (k, v) in stats.max_values.iter() {
+            match (k.as_str(), v) {
+                ("meta", ColumnValueStat::Column(map)) => {
+                    assert_eq!(2, map.len());
+
+                    let kafka = map.get("kafka").unwrap().as_column().unwrap();
+                    assert_eq!(3, kafka.len());
+                    let partition = kafka.get("partition").unwrap().as_value().unwrap();
+                    assert_eq!(1, partition.as_i64().unwrap());
+
+                    let producer = map.get("producer").unwrap().as_column().unwrap();
+                    assert_eq!(1, producer.len());
+                    let timestamp = producer.get("timestamp").unwrap().as_value().unwrap();
+                    assert_eq!("2021-06-22", timestamp.as_str().unwrap());
+                }
+                ("some_int", ColumnValueStat::Value(v)) => assert_eq!(400, v.as_i64().unwrap()),
+                ("some_bool", ColumnValueStat::Value(v)) => assert_eq!(true, v.as_bool().unwrap()),
+                ("some_string", ColumnValueStat::Value(v)) => {
+                    assert_eq!("PUT", v.as_str().unwrap())
+                }
+                ("date", ColumnValueStat::Value(v)) => {
+                    assert_eq!("2021-06-22", v.as_str().unwrap())
+                }
+                _ => assert!(false, "Key should not be present"),
+            }
+        }
+
+        // assert on null count
+        for (k, v) in stats.null_count.iter() {
+            match (k.as_str(), v) {
+                ("meta", ColumnCountStat::Column(map)) => {
+                    assert_eq!(2, map.len());
+
+                    let kafka = map.get("kafka").unwrap().as_column().unwrap();
+                    assert_eq!(3, kafka.len());
+                    let partition = kafka.get("partition").unwrap().as_value().unwrap();
+                    assert_eq!(0, partition);
+
+                    let producer = map.get("producer").unwrap().as_column().unwrap();
+                    assert_eq!(1, producer.len());
+                    let timestamp = producer.get("timestamp").unwrap().as_value().unwrap();
+                    assert_eq!(0, timestamp);
+                }
+                ("some_int", ColumnCountStat::Value(v)) => assert_eq!(100, *v),
+                ("some_bool", ColumnCountStat::Value(v)) => assert_eq!(100, *v),
+                ("some_string", ColumnCountStat::Value(v)) => assert_eq!(100, *v),
+                ("some_list", ColumnCountStat::Value(v)) => assert_eq!(100, *v),
+                ("some_nested_list", ColumnCountStat::Value(v)) => assert_eq!(0, *v),
+                ("date", ColumnCountStat::Value(v)) => assert_eq!(0, *v),
+                _ => assert!(false, "Key should not be present"),
+            }
+        }
+    }
+
+    async fn load_table(
+        table_uri: &str,
+        options: HashMap<String, String>,
+    ) -> Result<DeltaTable, DeltaTableError> {
+        let backend = crate::get_backend_for_uri_with_options(table_uri, options)?;
+        let mut table = DeltaTable::new(
+            table_uri,
+            backend,
+            crate::DeltaTableConfig {
+                require_tombstones: true,
+            },
+        )?;
+        table.load().await?;
+        Ok(table)
+    }
+
+    fn create_temp_table(table_path: &Path) {
+        let log_path = table_path.join("_delta_log");
+
+        let _ = std::fs::create_dir(log_path.as_path()).unwrap();
+        let _ = std::fs::write(
+            log_path.join("00000000000000000000.json"),
+            V0_COMMIT.as_str(),
+        )
+        .unwrap();
+    }
+
+    lazy_static! {
+        static ref SCHEMA: Value = json!({
+            "type": "struct",
+            "fields": [
+                {
+                    "name": "meta",
+                    "type": {
+                        "type": "struct",
+                        "fields": [
+                            {
+                                "name": "kafka",
+                                "type": {
+                                    "type": "struct",
+                                    "fields": [
+                                        {
+                                            "name": "topic",
+                                            "type": "string",
+                                            "nullable": true, "metadata": {}
+                                        },
+                                        {
+                                            "name": "partition",
+                                            "type": "integer",
+                                            "nullable": true, "metadata": {}
+                                        },
+                                        {
+                                            "name": "offset",
+                                            "type": "long",
+                                            "nullable": true, "metadata": {}
+                                        }
+                                    ],
+                                },
+                                "nullable": true, "metadata": {}
+                            },
+                            {
+                                "name": "producer",
+                                "type": {
+                                    "type": "struct",
+                                    "fields": [
+                                        {
+                                            "name": "timestamp",
+                                            "type": "string",
+                                            "nullable": true, "metadata": {}
+                                        }
+                                    ],
+                                },
+                                "nullable": true, "metadata": {}
+                            }
+                        ]
+                    },
+                    "nullable": true, "metadata": {}
+                },
+                { "name": "some_string", "type": "string", "nullable": true, "metadata": {} },
+                { "name": "some_int", "type": "integer", "nullable": true, "metadata": {} },
+                { "name": "some_bool", "type": "boolean", "nullable": true, "metadata": {} },
+                {
+                    "name": "some_list",
+                    "type": {
+                        "type": "array",
+                        "elementType": "string",
+                        "containsNull": true
+                    },
+                    "nullable": true, "metadata": {}
+                },
+                {
+                    "name": "some_nested_list",
+                    "type": {
+                        "type": "array",
+                        "elementType": {
+                            "type": "array",
+                            "elementType": "integer",
+                            "containsNull": true
+                        },
+                        "containsNull": true
+                    },
+                    "nullable": true, "metadata": {}
+               },
+               { "name": "date", "type": "string", "nullable": true, "metadata": {} },
+            ]
+        });
+        static ref V0_COMMIT: String = {
+            let schema_string = serde_json::to_string(&SCHEMA.clone()).unwrap();
+            let jsons = [
+                json!({
+                    "protocol":{"minReaderVersion":1,"minWriterVersion":2}
+                }),
+                json!({
+                    "metaData": {
+                        "id": "22ef18ba-191c-4c36-a606-3dad5cdf3830",
+                        "format": {
+                            "provider": "parquet", "options": {}
+                        },
+                        "schemaString": schema_string,
+                        "partitionColumns": ["date"], "configuration": {}, "createdTime": 1564524294376i64
+                    }
+                }),
+            ];
+
+            jsons
+                .iter()
+                .map(|j| serde_json::to_string(j).unwrap())
+                .collect::<Vec<String>>()
+                .join("\n")
+                .to_string()
+        };
+        static ref JSON_ROWS: Vec<Value> = {
+            std::iter::repeat(json!({
+                "meta": {
+                    "kafka": {
+                        "offset": 0,
+                        "partition": 0,
+                        "topic": "some_topic"
+                    },
+                    "producer": {
+                        "timestamp": "2021-06-22"
+                    },
+                },
+                "some_string": "GET",
+                "some_int": 302,
+                "some_bool": true,
+                "some_list": ["a", "b", "c"],
+                "some_nested_list": [[42], [84]],
+                "date": "2021-06-22",
+            }))
+            .take(100)
+            .chain(
+                std::iter::repeat(json!({
+                    "meta": {
+                        "kafka": {
+                            "offset": 100,
+                            "partition": 1,
+                            "topic": "another_topic"
+                        },
+                        "producer": {
+                            "timestamp": "2021-06-22"
+                        },
+                    },
+                    "some_string": "PUT",
+                    "some_int": 400,
+                    "some_bool": false,
+                    "some_list": ["x", "y", "z"],
+                    "some_nested_list": [[42], [84]],
+                    "date": "2021-06-22",
+                }))
+                .take(100),
+            )
+            .chain(
+                std::iter::repeat(json!({
+                    "meta": {
+                        "kafka": {
+                            "offset": 0,
+                            "partition": 0,
+                            "topic": "some_topic"
+                        },
+                        "producer": {
+                            "timestamp": "2021-06-22"
+                        },
+                    },
+                    "some_nested_list": [[42], null],
+                    "date": "2021-06-22",
+                }))
+                .take(100),
+            )
+            .collect()
+        };
+    }
+}
